+++ p2p/raintree/peers_manager_test.go
@@ -11,14 +11,15 @@ import (
 	"github.com/golang/mock/gomock"
 	"github.com/libp2p/go-libp2p/p2p/host/peerstore/pstoremem"
 	mocknet "github.com/libp2p/go-libp2p/p2p/net/mock"
-	"github.com/stretchr/testify/require"
-
 	"github.com/pokt-network/pocket/internal/testutil"
+	"github.com/pokt-network/pocket/p2p/config"
 	typesP2P "github.com/pokt-network/pocket/p2p/types"
 	mocksP2P "github.com/pokt-network/pocket/p2p/types/mocks"
 	"github.com/pokt-network/pocket/runtime/configs"
+	"github.com/pokt-network/pocket/runtime/defaults"
 	cryptoPocket "github.com/pokt-network/pocket/shared/crypto"
 	mockModules "github.com/pokt-network/pocket/shared/modules/mocks"
+	"github.com/stretchr/testify/require"
 )
 
 const (
@@ -96,11 +97,12 @@ func TestRainTree_Peerstore_HandleUpdate(t *testing.T) {
 			libp2pMockNet, err := mocknet.WithNPeers(1)
 			require.NoError(t, err)
 
-			rtCfg := &RainTreeConfig{
+			rtCfg := &config.RainTreeConfig{
 				Host:                  libp2pMockNet.Hosts()[0],
 				Addr:                  pubKey.Address(),
 				PeerstoreProvider:     pstoreProviderMock,
 				CurrentHeightProvider: currentHeightProviderMock,
+				MaxNonces:             defaults.DefaultP2PMaxNonces,
 			}
 
 			router, err := NewRainTreeRouter(mockBus, rtCfg)
@@ -163,7 +165,7 @@ func BenchmarkPeerstoreUpdates(b *testing.B) {
 			hostMock := mocksP2P.NewMockHost(ctrl)
 			hostMock.EXPECT().Peerstore().Return(libp2pPStore).AnyTimes()
 
-			rtCfg := &RainTreeConfig{
+			rtCfg := &config.RainTreeConfig{
 				Host:                  hostMock,
 				Addr:                  pubKey.Address(),
 				PeerstoreProvider:     pstoreProviderMock,
@@ -287,11 +289,12 @@ func testRainTreeMessageTargets(t *testing.T, expectedMsgProp *ExpectedRainTreeM
 	hostMock := mocksP2P.NewMockHost(ctrl)
 	hostMock.EXPECT().Peerstore().Return(libp2pPStore).AnyTimes()
 
-	rtCfg := &RainTreeConfig{
+	rtCfg := &config.RainTreeConfig{
 		Host:                  hostMock,
 		Addr:                  []byte{expectedMsgProp.orig},
 		PeerstoreProvider:     pstoreProviderMock,
 		CurrentHeightProvider: currentHeightProviderMock,
+		MaxNonces:             defaults.DefaultP2PMaxNonces,
 	}
 
 	router, err := NewRainTreeRouter(busMock, rtCfg)