+++ runtime/configs/config.go
@@ -6,7 +6,6 @@ import (
 	"strings"
 
 	"github.com/mitchellh/mapstructure"
-	// "github.com/pokt-network/pocket/logger"
 	"github.com/pokt-network/pocket/runtime/defaults"
 	"github.com/spf13/viper"
 )
@@ -129,9 +128,9 @@ func NewDefaultConfig(options ...func(*Config)) *Config {
 			BlockStorePath: defaults.DefaultPersistenceBlockStorePath,
 		},
 		P2P: &P2PConfig{
-			Port:            defaults.DefaultP2PPort,
-			ConnectionType:  defaults.DefaultP2PConnectionType,
-			MaxMempoolCount: defaults.DefaultP2PMaxMempoolCount,
+			Port:           defaults.DefaultP2PPort,
+			ConnectionType: defaults.DefaultP2PConnectionType,
+			MaxNonces:      defaults.DefaultP2PMaxNonces,
 		},
 		Telemetry: &TelemetryConfig{
 			Enabled:  defaults.DefaultTelemetryEnabled,