+++ runtime/docs/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.38] - 2023-05-08
+
+- Renamed `P2PConfig#MaxMempoolCount` to `P2PConfig#MaxNonces`
+- Renamed `DefaultP2PMaxMempoolCount` to `DefaultP2PMaxNonces`
+
 ## [0.0.0.37] - 2023-05-04
 
 - Add `network_id` field to the configs and give the default value of `"localnet"`